requirements.txt
cryptography~=42.0.7
fastapi~=0.111.0
pymongo~=4.7.2
pydantic~=2.7.1
PyJWT~=2.8.0